Rename hscard command
